ubuntu-gnome: Remove task to install packages
Almost all of these are installed by default in Ubuntu, even when picking the minimal installation option. I think I'd created this task because I'd migrated from Xubuntu and didn't want to install the full ubuntu-desktop package, but I should've instead created a one-off playbook similar to the upgrade-prep playbooks. Rather than maintain this, it's easier to remove; for example, 22.04 comes with gnome-shell-extension-desktop-icons-ng, not gnome-shell-extension-desktop-icons.

The only package that doesn't come with Ubuntu is gnome-shell-extension-system-monitor. It doesn't seem to be in the 22.04 repos yet anyway, so it will need some more work.